test(CheckBox): add demo test
